removed warning.
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/alugrid/trunk@969 0d966ed9-3843-0410-af09-ebfb50bd7c74

+++ src/parallel/mpAccess_MPI.cc
@@ -524,7 +524,7 @@ exchange (const vector < ObjectStream > & in,
     vout[l]. second = v[l].second; 
 
     // make sure buffer size is large enough
-    assert( out[l]._len >= vout[l]. second ); 
+    assert( (int) out[l]._len >= vout[l]. second ); 
   } 
   
   // exchange data
